move test with tables-definition to CommonPart2
add skip for Redshift
