+++ phpunit.xml.dist
@@ -65,6 +65,7 @@
 		<exclude>tests/Backend/CommonPart1/BigqueryTableExporterTest.php</exclude>
 		<exclude>tests/Backend/CommonPart1/BucketsTest.php</exclude>
 		<directory>tests/Backend/CommonPart2</directory>
+		<exclude>tests/Backend/CommonPart2/TablesDefinitionCommonTest.php</exclude>
 		<exclude>tests/Backend/Snowflake/CloneIntoWorkspaceTest.php</exclude>
 		<exclude>tests/Backend/Snowflake/CopyIntoWorkspaceTest.php</exclude>
 		<file>tests/Backend/Workspaces/LegacyWorkspacesLoadTest.php</file>
@@ -95,6 +96,7 @@
 		<exclude>tests/Backend/CommonPart1/BigqueryTableExporterTest.php</exclude>
 		<exclude>tests/Backend/CommonPart1/BucketsTest.php</exclude>
 		<directory>tests/Backend/CommonPart2</directory>
+		<exclude>tests/Backend/CommonPart2/TablesDefinitionCommonTest.php</exclude>
 		<exclude>tests/Backend/Snowflake/CloneIntoWorkspaceTest.php</exclude>
 		<exclude>tests/Backend/Snowflake/CopyIntoWorkspaceTest.php</exclude>
 		<file>tests/Backend/Workspaces/LegacyWorkspacesLoadTest.php</file>
@@ -121,6 +123,7 @@
 		<exclude>tests/Backend/CommonPart1/BucketsTest.php</exclude>
 		<exclude>tests/Backend/CommonPart1/TimeTravelTest.php</exclude>
 		<directory>tests/Backend/CommonPart2</directory>
+		<exclude>tests/Backend/CommonPart2/TablesDefinitionCommonTest.php</exclude>
 		<file>tests/Backend/Workspaces/LegacyWorkspacesLoadTest.php</file>
 		<file>tests/Backend/Workspaces/WorkspacesLoadTest.php</file>
 		<file>tests/Backend/Workspaces/WorkspacesUnloadTest.php</file>
@@ -195,6 +198,7 @@
 <!--		<file>tests/Backend/Workspaces/WorkspacesLoadTest.php</file>-->
 		<directory>tests/Backend/Synapse</directory>
 		<file>tests/Backend/Export/ExportParamsTest.php</file>
+		<file>tests/Backend/CommonPart2/TablesDefinitionCommonTest.php</file>
 
 		<!--FILE WORKSPACE-->
 		<file>tests/Backend/FileWorkspace/WorkspacesLoadTest.php</file>
@@ -213,6 +217,7 @@
 		<exclude>tests/Backend/CommonPart1/BucketsTest.php</exclude>
 		<exclude>tests/Backend/CommonPart1/TimeTravelTest.php</exclude>
 		<directory>tests/Backend/CommonPart2</directory>
+		<exclude>tests/Backend/CommonPart2/TablesDefinitionCommonTest.php</exclude>
 		<!--	excluded test is included in tests/Backend/Synapse/SynapseImportExportCommonTest.php -->
 		<exclude>tests/Backend/CommonPart1/ImportExportCommonTest.php</exclude>
 		<!--WORKSPACE-->
